requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.1
django-debug-toolbar==1.3.0
django-toolbelt==0.0.1
gunicorn==19.3.0
psycopg2==2.6
selenium==2.45.0
sqlparse==0.1.15
static3==0.6.0
whitenoise==1.0.6
